requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-data-tables
azure-servicebus
pyjwt
